third_party/blink/web_tests/external/wpt/fetch/private-network-access/service-worker-fetch-document-treat-as-public.tentative.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] treat-as-public to local: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] treat-as-public to private: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
Harness: the test ran to completion.
